+++ lib/Alchemy/Phrasea/SearchEngine/Elastic/Structure/ValueChecker.php
@@ -3,7 +3,7 @@
 namespace Alchemy\Phrasea\SearchEngine\Elastic\Structure;
 
 use Alchemy\Phrasea\SearchEngine\Elastic\FieldMapping;
-use Alchemy\Phrasea\SearchEngine\Elastic\Search\QueryHelper;
+use Alchemy\Phrasea\SearchEngine\Elastic\RecordHelper;
 use Assert\Assertion;
 
 class ValueChecker
@@ -19,7 +19,7 @@ class ValueChecker
     {
         Assertion::allIsInstanceOf($list, Typed::class);
         $is_numeric = is_numeric($value);
-        $is_valid_date = (QueryHelper::sanitizeDate($value) !== null);
+        $is_valid_date = (RecordHelper::sanitizeDate($value) !== null);
         $filtered = [];
         foreach ($list as $item) {
             switch ($item->getType()) {
